Fixed Home component test cases
