--- src/components/home/Home.tes.js
@@ -1,46 +1,0 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import toJSON from 'enzyme-to-json';
-import Home from './Home';
-import Footer from '../footer/Footer'
-import Header from '../header/Header'
-import Connections from '../connections/Connections'
-import findByTestAttr from '../../tests/utils';
-import  { UserContext } from '../../UserContext';
-
-
-const setUp = (props={}) => {  
-  const contextValues = {user: {}, setUser: () => {}};
-  jest.spyOn(React, 'useContext')
-      .mockImplementation(() => contextValues);
-  const wrapper = shallow(
-      <UserContext.Provider>
-        <Home />
-      </UserContext.Provider>
-  ).dive();
-  return wrapper;
-}
-
-describe('<Home /> component testing', () => {
-    let component;
-    beforeAll(() => {
-      component = setUp();
-    });
-    it('should render the component correctly', () => {
-         expect(component.exists()).toBe(true);
-    })
-    it('should render the <Home /> component correctly', () => {
-        console.log(component.debug());
-        const homeContainer = findByTestAttr(component, 'home-container');
-        expect(homeContainer.length).toBe(1);
-    });
-    it('should render <Header /> component correctly', () => {
-        expect(component.find('Header')).toHaveLength(1);
-    });
-    it('should render <Connections /> component correctly', () => {
-        expect(component.find('Connections')).toHaveLength(1);
-    });
-    it('should render <Footer /> component correctly', () => {
-       expect(component.find('Footer')).toHaveLength(1);
-    });
-});
